auth

+++ app/Http/Controllers/SessionController.php
@@ -6,6 +6,7 @@ use App\Models\Film;
 use App\Models\Hall;
 use App\Models\Session;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class SessionController extends Controller
 {
@@ -81,6 +82,11 @@ class SessionController extends Controller
      */
     public function destroy(string $id)
     {
+        if (! Gate::allows('destroy-session', Session::all()->where('id', $id)->first())) {
+            return redirect('/error')->with('message', 'У вас нет разрешения на удаление сенаса
+             номер ' . $id);
+        }
+
         Session::destroy($id);
         return redirect('/sessions');
     }
